Осталось доработать пользователя. "Редактирование" Все ок
