+++ db_crp/forms.py
@@ -36,7 +36,7 @@ class UserCreateForm(forms.Form):
 
 
 class UserEditForm(forms.Form):
-    """Форма для редактирования пользователя"""
+    """Редактирование пользователя"""
     email = forms.EmailField(label="Почта", required=False)
     password = forms.CharField(label="Пароль", widget=forms.PasswordInput, required=False)
     can_create_db = forms.BooleanField(label="Может создавать БД", required=False)
@@ -46,15 +46,6 @@ class UserEditForm(forms.Form):
     login = forms.BooleanField(label="Право входа", required=False)
     replication = forms.BooleanField(label="Право репликации", required=False)
     bypass_rls = forms.BooleanField(label="Bypass RLS", required=False)
-
-
-
-# class UserEditForm(forms.Form):
-#     """Редактирование пользователя"""
-#     email = forms.EmailField(label="Почта", required=False)
-#     password = forms.CharField(label="Пароль", widget=forms.PasswordInput, required=False)
-#     can_create_db = forms.BooleanField(label="Может создавать БД", required=False)
-#     is_superuser = forms.BooleanField(label="Суперпользователь", required=False)
 
 
 class GroupEditForm(forms.Form):
