+++ templates/send_email/send_user_edit.html
@@ -56,8 +56,6 @@
                 {% endfor %}
             </ul>
         {% endif %}
-        <p><strong>Может создавать БД:</strong> {{ can_create_db }}</p>
-        <p><strong>Суперпользователь:</strong> {{ is_superuser }}</p>
         <p>Если вы не запрашивали доступ к ресурсу, свяжитесь со службой поддержки.</p>
         <div class="footer">
             <p>С уважением, Администрация сайта.</p>
